==> 1_import_r.ipynb <==
## ropensci's JSTOR tool

The code below is based on [this tutorial](https://github.com/ropensci/jstor/blob/master/vignettes/analysing-n-grams.Rmd). It's used here to import metadata downloaded via JSTOR's Data for Research, to do some basic clean-up, and to export (via csv) to Python for analysis. 

install.packages("jstor")
install.packages("tidyverse")

library(jstor)
library(magrittr)
library(future)
library(tidyverse)

jst_preview_zip("receipt-id-1477091-part-001.zip")
#jst_preview_zip("data-improved.zip")

import_spec <- jst_define_import(article = jst_get_article)

The code below can take quite a long time to run depending on the size of the .zip file. 

jst_import_zip("receipt-id-1477091-part-001.zip", import_spec = import_spec, out_file = "part1")

imported_metadata <- c("part1_journal_article_jst_get_article-1.csv") %>% map_df(jst_re_import)

imported_metadata <- jst_augment(imported_metadata, quietly=TRUE)

### Go to notebook 2 >>
